refactor(secrets): rename keyvault package to secretsmanager
- Rename package from keyvault to secretsmanager to better reflect AWS service name
- Update import paths in run.go to reference new package name
- Move plugin and store implementations to new secretsmanager package
- No functional changes, purely organizational rename
